+++ test/python/tools/test_compiler.py
@@ -10,6 +10,7 @@
 """Compiler Test."""
 
 import unittest
+from unittest.mock import patch
 
 from qiskit import BasicAer
 from qiskit import QuantumRegister, ClassicalRegister, QuantumCircuit
@@ -23,7 +24,7 @@ from qiskit.converters import circuit_to_dag
 from qiskit.tools.qi.qi import random_unitary_matrix
 from qiskit.mapper.compiling import two_qubit_kak
 from qiskit.mapper.mapping import MapperError
-
+from qiskit.transpiler.passes import BarrierBeforeFinalMeasurements
 
 
 class TestCompiler(QiskitTestCase):
@@ -621,6 +622,27 @@ class TestCompiler(QiskitTestCase):
                     two_qubit_kak(unitary, verify_gate_sequence=True)
                 except MapperError as ex:
                     self.fail(str(ex))
+
+    barrier_pass = BarrierBeforeFinalMeasurements()
+
+    @patch.object(BarrierBeforeFinalMeasurements, 'run', wraps=barrier_pass.run)
+    def test_final_measurement_barrier_for_devices(self, mock_pass):
+        """Verify BarrierBeforeFinalMeasurements pass is called in default pipeline for devices."""
+
+        circ = QuantumCircuit.from_qasm_file(self._get_resource_path('example.qasm', Path.QASMS))
+        dag_circuit = circuit_to_dag(circ)
+        transpile_dag(dag_circuit, coupling_map=FakeRueschlikon().configuration().coupling_map)
+
+        self.assertTrue(mock_pass.called)
+
+    @patch.object(BarrierBeforeFinalMeasurements, 'run', wraps=barrier_pass.run)
+    def test_final_measurement_barrier_for_simulators(self, mock_pass):
+        """Verify BarrierBeforeFinalMeasurements pass is in default pipeline for simulators."""
+        circ = QuantumCircuit.from_qasm_file(self._get_resource_path('example.qasm', Path.QASMS))
+        dag_circuit = circuit_to_dag(circ)
+        transpile_dag(dag_circuit)
+
+        self.assertTrue(mock_pass.called)
 
 if __name__ == '__main__':
     unittest.main(verbosity=2)
